+++ src/include/lexer.hpp
@@ -92,7 +92,7 @@ struct Lexer {
 /// @param types A vector of strings that specify the types in the language
 /// @param comment What a single line comment is in the language
 std::unique_ptr<Lexer>
-lex_file(const char *src_code,
+lex_file(std::string &src_code,
          std::string fp,
          std::vector<std::string> &keywords,
          std::vector<std::string> &types,