+++ src/primitives/class.cpp
@@ -69,7 +69,8 @@ Class::type(void) const {
 std::shared_ptr<Obj>
 Class::binop(Token *op, std::shared_ptr<Obj> &other) {
     Err::err_wtok(op);
-    ERR(Err::Type::Fatal, "unable to perform binary operation on class type");
+    std::string msg = "unable to perform binary operation on class type";
+    throw InterpreterException(msg);
 }
 
 bool
@@ -106,13 +107,15 @@ void
 Class::spec_mutate(Token *op, const std::shared_ptr<Obj> &other) {
     (void)other;
     Err::err_wtok(op);
-    ERR_WARGS(Err::Type::Fatal, "invalid operator for special mutation `%s` on class type", op->lexeme().c_str());
+    std::string msg = "invalid operator for special mutation `"+op->lexeme()+"` on class type";
+    throw InterpreterException(msg);
 }
 
 std::shared_ptr<Obj>
 Class::unaryop(Token *op) {
     (void)op;
     Err::err_wtok(op);
-    ERR(Err::Type::Fatal, "invalid unary operator on class type");
+    std::string msg = "invalid unary operator on class type";
+    throw InterpreterException(msg);
     return nullptr; // unreachable
 }