+++ src/primitives/int.cpp
@@ -83,8 +83,10 @@ Int::binop(Token *op, std::shared_ptr<Obj> &other) {
             return std::make_shared<Int>(this->value() / dynamic_cast<Int *>(other.get())->value());
     } break;
     case TokenType::Percent: {
-        if (other->type() == Type::Float)
-            ERR(Err::Type::Fatal, "cannot use module `%%` with floating point");
+        if (other->type() == Type::Float) {
+            std::string msg = "cannot use module `%%` with floating point";
+            throw InterpreterException(msg);
+        }
         return std::make_shared<Int>(this->value() % dynamic_cast<Int *>(other.get())->value());
     } break;
     case TokenType::Lessthan: {
@@ -124,7 +126,8 @@ Int::binop(Token *op, std::shared_ptr<Obj> &other) {
     } break;
     default: {
         Err::err_wtok(op);
-        ERR(Err::Type::Fatal, "invalid binary operator");
+        std::string msg = "invalid binary operator";
+        throw InterpreterException(msg);
     }
     }
 }
@@ -183,7 +186,8 @@ Int::spec_mutate(Token *op, const std::shared_ptr<Obj> &other) {
     case TokenType::Percent_Equals: m_value %= prev; break;
     default: {
         Err::err_wtok(op);
-        ERR_WARGS(Err::Type::Fatal, "invalid operator for special mutation `%s`", op->lexeme().c_str());
+        std::string msg = "invalid operator for special mutation `"+op->lexeme()+"`";
+        throw InterpreterException(msg);
     } break;
     }
 }
@@ -195,7 +199,8 @@ Int::unaryop(Token *op) {
     case TokenType::Bang: return std::make_shared<Bool>(!m_value);
     default: {
         Err::err_wtok(op);
-        ERR(Err::Type::Fatal, "invalid unary operator on int type");
+        std::string msg = "invalid unary operator on int type";
+        throw InterpreterException(msg);
     }
     }
     return nullptr; // unreachable