+++ src/primitives/option.cpp
@@ -65,8 +65,10 @@ std::shared_ptr<Obj>
 Option::binop(Token *op, std::shared_ptr<Obj> &other) {
     ASSERT_BINOP_COMPAT(this, other.get(), op);
 
-    if (op->type() != TokenType::Double_Equals)
-        ERR(Err::Type::Fatal, "the only support binary operations on option types is equality `==`");
+    if (op->type() != TokenType::Double_Equals) {
+        std::string msg = "the only support binary operations on option types is equality `==`";
+        throw InterpreterException(msg);
+    }
 
     auto other2 = dynamic_cast<Option *>(other.get());
 
@@ -121,14 +123,16 @@ void
 Option::spec_mutate(Token *op, const std::shared_ptr<Obj> &other) {
     (void)other;
     Err::err_wtok(op);
-    ERR_WARGS(Err::Type::Fatal, "invalid operator for special mutation `%s` on option type", op->lexeme().c_str());
+    std::string msg = "invalid operator for special mutation `"+op->lexeme()+"` on option type";
+    throw InterpreterException(msg);
 }
 
 std::shared_ptr<Obj>
 Option::unaryop(Token *op) {
     (void)op;
     Err::err_wtok(op);
-    ERR(Err::Type::Fatal, "invalid unary operator on option type");
+    std::string msg = "invalid unary operator on option type";
+    throw InterpreterException(msg);
     return nullptr; // unreachable
 }